requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
Automat==24.8.1
backports.tarfile==1.2.0
binance-connector==3.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
exceptiongroup==1.2.2
frozenlist==1.5.0
gate-api==4.83.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.5.0
loguru==0.7.2
more-itertools==10.5.0
multidict==6.1.0
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
pybit==5.8.0
pycparser==2.22
pycryptodome==3.21.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-okx==0.3.3
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.1
Telethon==1.37.0
tomli==2.0.2
Twisted==24.10.0
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
yarl==1.17.1
zipp==3.20.2
zope.interface==7.1.1
